TAP14: update version number
Hopefully this much we can agree upon, at least ;)

Note that a parser is still technically allowed to interpret a `TAP
version 13` stream as TAP14 and be compliant with this spec, but this
behavior is not required.  (In other words, they can also throw that out
as an error, and also be compliant.)  Since the goal is to codify
existing behavior and add little to it, and all current "TAP13"
implementations allow `TAP version 13`, this follows logically.