clarify the syntax of the version line
A TAP14 stream *must* have a version line saying "TAP version 14" in
order to be TAP14. It's confusing to say that a TAP stream *must* have
that line when describing the syntax, though.

Reword this to make it clear that a version line must be that line, and
the semantics of *missing* that line is that this is not a TAP14 stream
-- not that it's a TAP14 stream with an error in it. Of course, a
harness that interprets TAP14, when given not-TAP14, can reasonably
decide to report an error.

Add the note about TAP13 compatibility into a new section on backwards
compatibility, making it clear that both the "may" handle TAP13 and the
"may" report an error for missing version, are... all about handling
backwards compatibility with That Which Is Not TAP14.

Ref. https://github.com/TestAnything/Specification/issues/41