upgrade harness support of older version streams from "may" to "should"
If a harness does not want to do so, they do not have to -- and the spec
is not changing that. But it's a good idea to do so, because TAP tries
hard to be compatible for a reason, and that reason implies that it's
safe and good to support it.

So, lean harder on the notion that it's a recommended move to do so.